app_deepr/requirements.txt
Flask
werkzeug
itsdangerous
flask-sqlalchemy
psycopg2-binary
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-sqlite3
opentelemetry-instrumentation-logging
opentelemetry-instrumentation-psycopg2
opentelemetry-instrumentation-requests